requirements.txt
allure-pytest==2.8.9
pytest==5.3.5
pytest-xdist==1.31.0
selenium==3.141.0
webdriver-manager==2.3.0
